+++ projects/hipcub/benchmark/benchmark_warp_merge_sort.cpp
@@ -493,7 +493,7 @@ void add_benchmarks(primbench::executor& executor,
     BENCHMARK_TYPE(int, 256);
     BENCHMARK_TYPE(int8_t, 256);
     BENCHMARK_TYPE(uint8_t, 256);
-    BENCHMARK_TYPE(long long, 256);
+    BENCHMARK_TYPE(int64_t, 256);
 }
 
 int main(int argc, char* argv[])
